change spoil view to qualityView and is a component
