+++ app/templates/components/quality-view.hbs
@@ -1,0 +1,24 @@
+<td style="background: {{concat "rgba(56,112,255," truthyColor ")"}}">
+  {{#ui-popup html=(concat
+    '<table class="colored-note-details">
+    <tr><td>Spoil </td><td> ' (if skill.productionTemplate.spoil skill.productionTemplate.spoil "Non testé") '</td></tr>
+    <tr><td>Responsive </td><td> ' (if skill.productionTemplate.responsive skill.productionTemplate.responsive "Non testé") '</td></tr>
+    <tr><td>Non/Mal voyant </td><td> ' (if skill.productionTemplate.accessibility1 skill.productionTemplate.accessibility1 "Non testé") '</td></tr>
+    <tr><td>Daltonien </td><td> ' (if skill.productionTemplate.accessibility2 skill.productionTemplate.accessibility2 "Non testé") '</td></tr>
+    <tr><td>Indice </td><td> ' (if skill.clueStatus skill.clueStatus "Pas d'indice") '</td></tr>
+    <tr><td>Tuto comprendre </td><td> ' skill.tutoSolutionCount '</td></tr>
+    <tr><td>Tuto en savoir + </td><td> ' skill.tutoMoreCount '</td></tr>
+    '(if skill.productionTemplate.timer (concat '<tr><td>Timer</td><td>' skill.productionTemplate.timer '</td></tr>') "") '
+    </table>')}}
+    <div class="skill-name {{skill.descriptionCSS}} colored-note">
+      <span>{{skill.name}}</span>
+      <span>{{truthyColor}}</span>
+    </div>
+    {{#if haveTutorial}}
+      <i class="lab icon {{haveTutorial}}"></i>
+    {{/if}}
+    {{#if skill.productionTemplate.timer }}
+      <i class="time icon"></i>
+    {{/if}}
+  {{/ui-popup}}
+</td>
